+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ai-sdk/azure
 
+## 4.0.31
+
+### Patch Changes
+
+- Updated dependencies [1bec07d]
+  - @ai-sdk/provider-utils@5.0.21
+  - @ai-sdk/openai@4.0.30
+  - @ai-sdk/deepseek@3.0.22
+
+## 4.0.30
+
+### Patch Changes
+
+- Updated dependencies [160ccdb]
+  - @ai-sdk/provider-utils@5.0.20
+  - @ai-sdk/deepseek@3.0.21
+  - @ai-sdk/openai@4.0.29
+
+## 4.0.29
+
+### Patch Changes
+
+- Updated dependencies [79e133c]
+  - @ai-sdk/provider@4.0.5
+  - @ai-sdk/deepseek@3.0.20
+  - @ai-sdk/openai@4.0.28
+  - @ai-sdk/provider-utils@5.0.19
+
 ## 4.0.28
 
 ### Patch Changes