+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,39 @@
 
 ### Patch Changes
 
+- Updated dependencies [3836a85]
+  - ai@7.0.52
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@7.0.51
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [79e133c]
+- Updated dependencies [da64b51]
+  - ai@7.0.50
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@7.0.49
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@7.0.48
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@7.0.47
 
 ## 0.0.1