+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/valibot
 
+## 3.0.21
+
+### Patch Changes
+
+- Updated dependencies [1bec07d]
+  - @ai-sdk/provider-utils@5.0.21
+
+## 3.0.20
+
+### Patch Changes
+
+- Updated dependencies [160ccdb]
+  - @ai-sdk/provider-utils@5.0.20
+
+## 3.0.19
+
+### Patch Changes
+
+- @ai-sdk/provider-utils@5.0.19
+
 ## 3.0.18
 
 ### Patch Changes